Reorganize modules related to layout service - move them into their own folder and separate rest/graphql, similar to what we did with dictionary modules
