+++ packages/sitecore-jss/src/layout/layout-service.ts
@@ -1,0 +1,19 @@
+import { LayoutServiceData } from './models';
+import { IncomingMessage, ServerResponse } from 'http';
+
+export interface LayoutService {
+  /**
+   * Fetch layout data for an item.
+   * @param {string} itemPath
+   * @param {string} [language]
+   * @param {IncomingMessage} [req] Request instance
+   * @param {ServerResponse} [res] Response instance
+   * @returns {Promise<LayoutServiceData>} layout data
+   */
+  fetchLayoutData(
+    itemPath: string,
+    language?: string,
+    req?: IncomingMessage,
+    res?: ServerResponse
+  ): Promise<LayoutServiceData>;
+}
