+++ packages/sitecore-jss/src/layout/utils.ts
@@ -1,4 +1,4 @@
-import { ComponentRendering, Field, HtmlElementRendering, Item } from './data-models';
+import { ComponentRendering, Field, HtmlElementRendering, Item } from './models';
 
 type Fields = { [name: string]: Field | Item[] };
